+++ src/Model/PropModel.php
@@ -4,10 +4,36 @@ declare(strict_types=1);
 
 namespace App\Model;
 
+use App\Entity\Attr;
+
 class PropModel
 {
+    private int $id;
+    private int $deviceId;
     private ?string $name = null;
-    private ?string $attrUrl = null;
+    private ?array $link;
+
+    public function getId(): int
+    {
+        return $this->id;
+    }
+
+    public function setId(int $id): static
+    {
+        $this->id = $id;
+        return $this;
+    }
+
+    public function getDeviceId(): int
+    {
+        return $this->deviceId;
+    }
+
+    public function setDeviceId(int $deviceId): static
+    {
+        $this->deviceId = $deviceId;
+        return $this;
+    }
 
     public function getName(): ?string
     {
@@ -20,15 +46,19 @@ class PropModel
         return $this;
     }
 
-    public function getAttrUrl(): ?string
+    /* 
+    public function getLinks(): ?array
     {
-        return $this->attrUrl;
+        return $this->link;
     }
 
-    public function setAttrUrl(int $id): static
+    public function setLinks(Attr $attr): static
     {
-        $this->attrUrl = '/api/attr/' . $id;
+        $this->link = [
+            'self' => (string) '/api/devices/property/' . $this->getId(),
+            'href' => (string) '/api/devices/property/attribut/' . $attr->getId()
+        ];
         return $this;
-    }
+    }*/
 
 }