+++ src/Service/AttrService.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\Service;
 
+use App\Entity\Attr;
 use App\Entity\Device;
 use App\Model\AttrModel;
 use App\Mapper\AttrMapper;
@@ -26,14 +27,23 @@ class AttrService
 
     public function getAll(): Collection|array
     {
-        $attrModel = $this->attrRepo->findAll();
-        return $this->attrMapper->EntitiesToModels($attrModel);
+        return $this->attrRepo->findAll();
     }
 
-    public function getAttr(int $id): AttrModel
+    public function getAllModel(): Collection|array
+    {
+        return $this->attrMapper->entitiesToModels($this->getAll());
+    }
+
+    public function getAttrById(int $id): Attr
+    {
+        return $this->attrRepo->findOneById($id);
+    }
+
+    public function getAttrModelById(int $id): AttrModel
     {
         $deviceEntity = $this->attrRepo->findOneById($id);
-        return $this->attrMapper->EntityToModel($deviceEntity);
+        return $this->attrMapper->entityToModel($this->getAttrById($id));
     }
 
 }