+++ src/Service/ClientService.php
@@ -5,9 +5,10 @@ declare(strict_types=1);
 namespace App\Service;
 
 use App\Entity\Client;
-use App\Model\ClientModel;
-use App\Mapper\ClientMapper;
-use App\Mapper\ClientsMapper;
+use App\Model\ClientDetailsModel;
+use App\Model\ClientSummaryModel;
+use App\Mapper\ClientDetailsMapper;
+use App\Mapper\ClientSummaryMapper;
 use App\Repository\ClientRepository;
 use Doctrine\Common\Collections\Collection;
 
@@ -16,22 +17,10 @@ class ClientService
 
     public function __construct(
         private ClientRepository $clientRepo,
-        private ClientMapper $clientMapper,
-        private ClientsMapper $clientsMapper,
+        private ClientDetailsMapper $clientDetailsMapper,
+        private ClientSummaryMapper $clientSummaryMapper,
     ) {
 
-    }
-
-    public function getClients(): Collection|array
-    {
-        $clientsModel = $this->clientRepo->findAll();
-        return $this->clientsMapper->EntitiesToModels($clientsModel);
-    }
-
-    public function getClient(int $id): ClientModel|null
-    {
-        $clientModel = $this->clientRepo->findOneById($id);
-        return $this->clientMapper->EntityToModel($clientModel);
     }
 
     public function getById(int $id): Client|null
@@ -39,9 +28,34 @@ class ClientService
         return $this->clientRepo->findOneById($id);
     }
 
-    public function getFirstId(): int|null
+    public function getAll(): array|null
     {
-        return $this->clientRepo->findFirstClientId();
+        return $this->clientRepo->findAll();
+    }
+
+    public function getClientById(int $id): Client|null
+    {
+        return $this->clientRepo->findOneById($id);
+    }
+
+    public function getUsersByClientId(int $id): Collection|array
+    {
+        return $this->clientRepo->findByClientId($id);
+    }
+
+    public function getClientsList(): Collection|array
+    {
+        return $this->clientSummaryMapper->entitiesToModels($this->getAll());
+    }
+
+    public function getClientSummary(int $id): ClientSummaryModel|null
+    {
+        return $this->clientSummaryMapper->entityToModel($this->getClientById($id));
+    }
+
+    public function getClientDetails(Client $client): ClientDetailsModel|null
+    {
+        return $this->clientDetailsMapper->entityToModel($client);
     }
 
 }